cambiar tipo de dato de monto a BigDecimal

+++ src/main/java/com/example/leerDatos/entitys/ResumenDto.java
@@ -3,16 +3,17 @@ package com.example.leerDatos.entitys;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 @Getter @Setter
 public class ResumenDto {
 
-    private int montoTotal;
+    private BigDecimal montoTotal;
     private int cantidadClientes;
     private Set<String> separarPorTipoMoneda;
-    private int montoMaximo;
-    private int montoMinimo;
+    private BigDecimal montoMaximo;
+    private BigDecimal montoMinimo;
     private Set<String> tipo;
     private Set<String> categoria;
 
@@ -20,7 +21,7 @@ public class ResumenDto {
     public ResumenDto() {
     }
 
-    public ResumenDto(int montoTotal, int cantidadClientes, Set<String> separarPorTipoMoneda, int montoMaximo, int montoMinimo) {
+    public ResumenDto(BigDecimal montoTotal, int cantidadClientes, Set<String> separarPorTipoMoneda, BigDecimal montoMaximo, BigDecimal montoMinimo) {
         this.montoTotal = montoTotal;
         this.cantidadClientes = cantidadClientes;
         this.separarPorTipoMoneda = separarPorTipoMoneda;
@@ -28,7 +29,7 @@ public class ResumenDto {
         this.montoMinimo = montoMinimo;
     }
 
-    public ResumenDto(int montoTotal, int cantidadClientes, Set<String> separarPorTipoMoneda, int montoMaximo, int montoMinimo,Set<String> tipo, Set<String> categoria) {
+    public ResumenDto(BigDecimal montoTotal, int cantidadClientes, Set<String> separarPorTipoMoneda, BigDecimal montoMaximo, BigDecimal montoMinimo,Set<String> tipo, Set<String> categoria) {
         this(montoTotal,cantidadClientes,separarPorTipoMoneda,montoMaximo,montoMinimo);
         this.tipo = tipo;
         this.categoria = categoria;
